不要な環境変数を削除
