+++ api/config/config_test.go
@@ -11,19 +11,12 @@ func TestNew(t *testing.T) {
 		want    *Config
 		wantErr bool
 	}{
-		{name: "ケース１",want: &Config{Env: "dev",Port: "8080",DBHost: "dbsrv01",DBUser: "mh-api",DBPassword: "password", DBName: "mh-api",USER: "admin",PASSWORD: "password",SECRET_KEY: "secret_key",TOKEN_LIFETIME: "3600"},wantErr: false},
+		{name: "正常系",want: &Config{Env: "dev",Port: "8080",Database_url: "user:password@tcp(localhost)/test?charset=utf8&parseTime=True&loc=Local&tls=true"},wantErr: false},
 	}
 	for _, tt := range tests {
 		t.Setenv("ENV",tt.want.Env)
 		t.Setenv("PORT",tt.want.Port)
-		t.Setenv("DB_HOST",tt.want.DBHost)
-		t.Setenv("DB_USER",tt.want.DBUser)
-		t.Setenv("DB_PASSWORD",tt.want.DBPassword)
-		t.Setenv("DB_NAME",tt.want.DBName)
-		t.Setenv("USER",tt.want.USER)
-		t.Setenv("PASSWORD",tt.want.PASSWORD)
-		t.Setenv("SECRET_KEY",tt.want.SECRET_KEY)
-		t.Setenv("TOKEN_LIFETIME",tt.want.TOKEN_LIFETIME)
+		t.Setenv("DATABASE_URL",tt.want.Database_url)
 
 		t.Run(tt.name, func(t *testing.T) {
 			got, err := New()
