--- api/controller/middleware/middleware.go
@@ -1,91 +1,0 @@
-package middleware
-
-import (
-	"fmt"
-	"mh-api/api/config"
-	"net/http"
-	"strconv"
-	"time"
-
-	"github.com/gin-contrib/cors"
-	"github.com/gin-gonic/gin"
-	"github.com/golang-jwt/jwt/v5"
-)
-
-func AuthMiddleware(ctx *gin.Context) {
-	cfg, err := config.New()
-	if err != nil {
-		panic(err)
-	}
-	tokenString, err := ctx.Cookie("token")
-	if err != nil {
-		ctx.JSON(http.StatusUnauthorized, gin.H{
-			"message": "Unauthorized",
-		})
-		ctx.Abort()
-		return
-	}
-
-	_, err = parseToken(tokenString,cfg)
-	if err != nil {
-		ctx.JSON(http.StatusUnauthorized, gin.H{
-			"message": "Invalid token",
-		})
-		ctx.Abort()
-		return
-	}
-
-	ctx.Next()
-}
-
-func CORS() gin.HandlerFunc {
-	return cors.New(cors.Config{
-		AllowOrigins: []string{
-			"*",
-		},
-		AllowMethods: []string{
-			"POST",
-			"GET",
-			"OPTIONS",
-		},
-		AllowHeaders: []string{
-			"Content-Type",
-		},
-		AllowCredentials: false,
-		MaxAge: 24 * time.Hour,
-	})
-}
-
-func GenerateToken(userId string,cfg *config.Config) (string, error) {
-	secretKey := cfg.SECRET_KEY// 暗号化、復号化するためのキー
-	tokenLifeTime, err := strconv.Atoi(cfg.TOKEN_LIFETIME)
-	if err != nil {
-		return "", err
-	}
-
-	claims := jwt.MapClaims{
-		"user_id": userId,
-		"exp":     time.Now().Add(time.Hour * time.Duration(tokenLifeTime)).Unix(),
-	}
-	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
-	tokenString, err := token.SignedString([]byte(secretKey))
-	if err != nil {
-		return "", err
-	}
-
-	return tokenString, nil
-}
-
-func parseToken(tokenString string,cfg *config.Config) (*jwt.Token, error) {
-	token, err := jwt.Parse(tokenString, func(token *jwt.Token) (interface{}, error) {
-		if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
-			return nil, fmt.Errorf("unexpected signing method: %v", token.Header["alg"])
-		}
-		secret_key := cfg.SECRET_KEY
-		return []byte(secret_key), nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	return token, nil
-}
